requirements.txt
Flask==0.9
Flask-Login==0.1.3
Flask-OAuth==0.12
Jinja2==2.6
Werkzeug==0.8.3
httplib2==0.8
itsdangerous==0.17
oauth2==1.5.211
pymongo==2.5
pysha3==0.3
requests==1.1.0
wsgiref==0.1.2
